rename: rename package pieces to piece
