+++ src/main/java/domain/piece/Soldier.java
@@ -1,4 +1,4 @@
-package domain.pieces;
+package domain.piece;
 
 import domain.*;
