+++ src/test/java/BoardTest.java
@@ -3,7 +3,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import domain.Board;
 import domain.Camp;
 import domain.Position;
-import domain.pieces.Soldier;
+import domain.piece.Soldier;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
